% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{check_docker_image}
\alias{check_docker_image}
\title{Check if HybPhaser Docker Image Exists}
\usage{
check_docker_image(image = "joelnitta/hybphaser:latest", pull = FALSE)
}
\arguments{
\item{image}{Name of the Docker image. Default is
"joelnitta/hybphaser:latest"}

\item{pull}{Logical; if TRUE and image is not found, attempt to pull it}
}
\value{
Logical; TRUE if image exists,  FALSE otherwise
}
\description{
Verifies that the HybPhaser Docker image is available locally
}
\examples{
\dontrun{
check_docker_image("joelnitta/hybphaser:latest")
}
}
