% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_hybpiper.R
\name{check_hybpiper_conda}
\alias{check_hybpiper_conda}
\title{Check if HybPiper is Available in a Conda Environment}
\usage{
check_hybpiper_conda(env_name = "hybpiper_env", quiet = FALSE)
}
\arguments{
\item{env_name}{Character; conda environment name}

\item{quiet}{Logical; if TRUE, suppresses messages}
}
\value{
Logical; TRUE if HybPiper is available in the environment
}
\description{
Verifies that \code{hybpiper} can be executed from the specified conda
environment.
}
