% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_bbsplit_results.R
\name{collate_bbsplit_results_from_config}
\alias{collate_bbsplit_results_from_config}
\title{Collate BBSplit Results Using a HybPhaser Config File}
\usage{
collate_bbsplit_results_from_config(config_file = "./config.txt")
}
\arguments{
\item{config_file}{Path to HybPhaser configuration file.}
}
\value{
Invisibly returns the same object as \code{collate_bbsplit_results()}.
}
\description{
Reads required settings from \code{config.txt} and runs
\code{collate_bbsplit_results()}.
}
