% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_phasing_stats.R
\name{collate_phasing_stats_from_config}
\alias{collate_phasing_stats_from_config}
\title{Collate Phasing BBSplit Stats Using a HybPhaser Config File}
\usage{
collate_phasing_stats_from_config(config_file = "./config.txt")
}
\arguments{
\item{config_file}{Path to HybPhaser configuration file.}
}
\value{
Invisibly returns the same object as \code{collate_phasing_stats()}.
}
\description{
Reads required settings from \code{config.txt} and runs
\code{collate_phasing_stats()}.
}
