% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{create_test_namelist}
\alias{create_test_namelist}
\title{Create Test Namelist File}
\usage{
create_test_namelist(file_path, samples = c("sample1", "sample2"))
}
\arguments{
\item{file_path}{Path where namelist should be created}

\item{samples}{Character vector of sample names}
}
\value{
Path to created file (invisibly)
}
\description{
Creates a namelist file for testing
}
\examples{
\dontrun{
namelist <- create_test_namelist(
  file.path(tempdir(), "namelist.txt"),
  samples = c("sample1", "sample2")
)
}
}
