% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{create_test_targets}
\alias{create_test_targets}
\title{Create Test Target File}
\usage{
create_test_targets(
  file_path,
  genes = c("gene001", "gene002"),
  seq_length = 150
)
}
\arguments{
\item{file_path}{Path where target file should be created}

\item{genes}{Character vector of gene names}

\item{seq_length}{Integer; length of sequences to generate}
}
\value{
Path to created file (invisibly)
}
\description{
Creates a FASTA file with target sequences (baits) for testing
}
\examples{
\dontrun{
target_file <- create_test_targets(
  file.path(tempdir(), "targets.fasta")
)
}
}
