% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{.run_docker}
\alias{.run_docker}
\title{Run a Command in the rhybphaser Docker Container}
\usage{
.run_docker(cmd, volumes = NULL, image = "joelnitta/rhybphaser:latest", ...)
}
\arguments{
\item{cmd}{Character vector; command to run}

\item{volumes}{Named character vector; host paths to mount (names are
container paths)}

\item{image}{Docker image name}

\item{...}{Additional arguments passed to system2}
}
\value{
Exit code from Docker command
}
\description{
Internal function to execute commands in the rhybphaser Docker container
}
\keyword{internal}
