% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{hybphaser_example}
\alias{hybphaser_example}
\title{Get Path to HybPhaser Example Data}
\usage{
hybphaser_example(file_name = NULL)
}
\arguments{
\item{file_name}{Name of the example file. If NULL, returns the extdata
directory. Available files:
\itemize{
\item "clade_references.csv"
\item "phasing_prep.csv"
}}
}
\value{
Character path to the example file or extdata directory
}
\description{
Returns the full path to example data files included with the package
}
\examples{
# Get path to extdata directory
data_dir <- hybphaser_example()

# Get path to specific example file
clade_ref <- hybphaser_example("clade_references.csv")
}
