% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phased_hybpiper_helpers.R
\name{make_phased_namelist}
\alias{make_phased_namelist}
\title{Create Namelist from Phased Read Files}
\usage{
make_phased_namelist(
  phased_reads_dir,
  phased_hp_dir,
  output_filename = "namelist_phased.txt"
)
}
\arguments{
\item{phased_reads_dir}{Path to directory containing phased read files.}

\item{phased_hp_dir}{Directory where the namelist file should be written.}

\item{output_filename}{Name of output namelist file.}
}
\value{
Character scalar. Path to created namelist file.
}
\description{
Creates a namelist text file from phased read file names in a directory.
}
\examples{
\dontrun{
namelist_path <- make_phased_namelist(
  phased_reads_dir = "hybphaser_output/05_phasing/phased_reads",
  phased_hp_dir = "hybpiper_output_phased"
)
}
}
