% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_phasing_script.R
\name{prepare_phasing_script_from_config}
\alias{prepare_phasing_script_from_config}
\title{Prepare Phasing Script Using a HybPhaser Config File}
\usage{
prepare_phasing_script_from_config(config_file = "./config.txt")
}
\arguments{
\item{config_file}{Path to HybPhaser configuration file.}
}
\value{
Invisibly returns the same object as \code{prepare_phasing_script()}.
}
\description{
Reads required phasing settings from \code{config.txt} and runs
\code{prepare_phasing_script()}.
}
