% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhybphaser-package.R
\docType{package}
\name{rhybphaser-package}
\alias{rhybphaser}
\alias{rhybphaser-package}
\title{rhybphaser: Phasing of Sequence Reads for Polyploid Hybrids from Target Capture Data}
\description{
An R implementation of the HybPhaser method by Nauheimer et al. (original code at \url{https://github.com/LarsNauheimer/HybPhaser}). HybPhaser detects hybrids by measuring heterozygosity in target capture datasets and phases hybrid accessions by separating reads according to similarity with selected taxa that represent parental clades. It is designed as an extension to the HybPiper assembly pipeline.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joelnitta/rhybphaser}
  \item Report bugs at \url{https://github.com/joelnitta/rhybphaser/issues}
}

}
\author{
\strong{Maintainer}: Joel Nitta \email{joelnitta@gmail.com} (\href{https://orcid.org/0000-0003-4719-7472}{ORCID})

Authors:
\itemize{
  \item Joel Nitta \email{joelnitta@gmail.com} (\href{https://orcid.org/0000-0003-4719-7472}{ORCID})
}

}
\keyword{internal}
