% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_phasing_script.R
\name{run_phasing_from_config}
\alias{run_phasing_from_config}
\title{Run BBSplit Phasing Using a HybPhaser Config File}
\usage{
run_phasing_from_config(config_file = "./config.txt")
}
\arguments{
\item{config_file}{Path to HybPhaser configuration file.}
}
\value{
Invisibly returns the same object as \code{run_phasing()}.
}
\description{
Reads required phasing settings from \code{config.txt} and runs
\code{run_phasing()}.
}
